requirements.txt
boto3==1.26.133
DataProfiler==0.8.9
duckdb==0.8.1
numpy==1.24.2
openai==0.27.6
pandas==1.5.3
Pillow==9.5.0
plotly==5.4.0
psutil==5.9.0
psycopg2==2.9.2
Requests==2.31.0
streamlit==1.22.0
